+++ apps/server/src/lib.rs
@@ -1,5 +1,6 @@
 mod db;
 mod graphql;
+mod query;
 
 pub fn add(left: usize, right: usize) -> usize {
     left + right
